translate dot product to go assembly
