+++ enterprise/internal/embeddings/dot_noamd64.go
@@ -1,4 +1,4 @@
-//go:build !(amd64 && cgo)
+//go:build !amd64
 
 package embeddings
